Refactor and add unit tests to API interceptors (#3647)
