+++ .github/workflows/client-build.yml
@@ -94,11 +94,12 @@ jobs:
 
       - name: Run the jest tests
         if: github.event_name == 'pull_request'
-        uses: anuraag016/Jest-Coverage-Diff@V1.1
+        uses: hetunandu/Jest-Coverage-Diff@feature/improve-comment-ui
         with:
           fullCoverageDiff: false
           runCommand: cd app/client && REACT_APP_ENVIRONMENT=${{steps.vars.outputs.REACT_APP_ENVIRONMENT}} yarn run test:unit
-          delta: 0.5
+          # percentage of drop in coverage accepted
+          delta: 2
 
       # We burn React environment & the Segment analytics key into the build itself.
       # This is to ensure that we don't need to configure it in each installation
