+++ app/client/.gitignore
@@ -8,6 +8,7 @@
 # testing
 /coverage
 
+
 # production
 /build
