+++ app/client/src/api/ApplicationApi.tsx
@@ -1,4 +1,4 @@
-import Api from "./Api";
+import Api from "api/Api";
 import { ApiResponse } from "./ApiResponses";
 import { AxiosPromise } from "axios";
 import { AppColorCode } from "constants/DefaultTheme";
