+++ app/client/src/api/OrgApi.ts
@@ -1,5 +1,5 @@
 import { AxiosPromise } from "axios";
-import Api from "./Api";
+import Api from "api/Api";
 import { ApiResponse } from "./ApiResponses";
 import { OrgRole, Org } from "constants/orgConstants";
