+++ app/client/src/api/UserApi.tsx
@@ -1,5 +1,5 @@
 import { AxiosPromise } from "axios";
-import Api from "./Api";
+import Api from "api/Api";
 import { ApiResponse } from "./ApiResponses";
 
 export interface LoginUserRequest {
