+++ app/client/src/sagas/ActionExecutionSagas.ts
@@ -49,7 +49,7 @@ import {
 } from "actions/actionActions";
 import { Action } from "entities/Action";
 import ActionAPI, {
-  ActionApiResponse,
+  ActionExecutionResponse,
   ActionResponse,
   ExecuteActionRequest,
   PaginationField,
@@ -328,12 +328,12 @@ export const getActionTimeout = (
   return undefined;
 };
 const createActionExecutionResponse = (
-  response: ActionApiResponse,
+  response: ActionExecutionResponse,
 ): ActionResponse => ({
   ...response.data,
   ...response.clientMeta,
 });
-const isErrorResponse = (response: ActionApiResponse) => {
+const isErrorResponse = (response: ActionExecutionResponse) => {
   return !response.data.isExecutionSuccess;
 };
 
@@ -456,7 +456,7 @@ export function* executeActionSaga(
       viewMode: appMode === APP_MODE.PUBLISHED,
     };
     const timeout = yield select(getActionTimeout, actionId);
-    const response: ActionApiResponse = yield ActionAPI.executeAction(
+    const response: ActionExecutionResponse = yield ActionAPI.executeAction(
       executeActionRequest,
       timeout,
     );
@@ -667,7 +667,7 @@ function* runActionSaga(
     const timeout = yield select(getActionTimeout, actionId);
     const appMode = yield select(getAppMode);
     const viewMode = appMode === APP_MODE.PUBLISHED;
-    const response: ActionApiResponse = yield ActionAPI.executeAction(
+    const response: ActionExecutionResponse = yield ActionAPI.executeAction(
       {
         actionId,
         params,
@@ -773,7 +773,7 @@ function* executePageLoadAction(pageAction: PageAction) {
       appName: currentApp.name,
       isExampleApp: currentApp.appIsExample,
     });
-    const response: ActionApiResponse = yield ActionAPI.executeAction(
+    const response: ActionExecutionResponse = yield ActionAPI.executeAction(
       executeActionRequest,
       pageAction.timeoutInMillisecond,
     );
